--- extensions/placeholder-data-source/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -1,1 +1,0 @@
-de.sovity.edc.extension.placeholderdatasource.PlaceholderDataSourceExtension
