+++ extensions/sovity-edc-extensions-package/build.gradle.kts
@@ -6,7 +6,6 @@ plugins {
 
 dependencies {
     // Policies
-    api(project(":extensions:placeholder-data-source"))
     api(project(":extensions:policy-referring-connector"))
     api(project(":extensions:policy-time-interval"))
     api(project(":extensions:policy-always-true"))
