+++ extensions/wrapper/wrapper-common-mappers/src/test/java/de/sovity/edc/ext/wrapper/api/common/mappers/Factory.java
@@ -24,7 +24,6 @@ import de.sovity.edc.ext.wrapper.api.common.mappers.asset.utils.ShortDescription
 import de.sovity.edc.ext.wrapper.api.common.mappers.dataaddress.DataSourceMapper;
 import de.sovity.edc.ext.wrapper.api.common.mappers.dataaddress.http.HttpDataSourceMapper;
 import de.sovity.edc.ext.wrapper.api.common.mappers.dataaddress.http.HttpHeaderMapper;
-import de.sovity.edc.extension.placeholderdatasource.PlaceholderEndpointService;
 import org.eclipse.edc.jsonld.TitaniumJsonLd;
 import org.eclipse.edc.spi.monitor.Monitor;
 import org.eclipse.edc.transform.spi.TypeTransformerRegistry;
