+++ settings.gradle.kts
@@ -6,7 +6,6 @@ include(":extensions:catalog-crawler:catalog-crawler-launcher-base")
 include(":extensions:catalog-crawler:catalog-crawler-e2e-test")
 include(":extensions:contract-termination")
 include(":extensions:database-direct-access")
-include(":extensions:placeholder-data-source")
 include(":extensions:edc-ui-config")
 include(":extensions:last-commit-info")
 include(":extensions:policy-always-true")
